+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/OrmImageExport.cs
@@ -271,19 +271,10 @@ namespace GLTFast.Export
             var exportTexture = new Texture2D(
                 width,
                 height,
-#if UNITY_2023_2_OR_NEWER
-                // ~20 times faster texture construction
                 SystemInfo.IsFormatSupported(GraphicsFormat.R8G8B8_UNorm, GraphicsFormatUsage.Sample)
                     ? GraphicsFormat.R8G8B8_UNorm
                     : GraphicsFormat.R8G8B8A8_UNorm,
                 TextureCreationFlags.DontInitializePixels | TextureCreationFlags.DontUploadUponCreate
-#else
-                // ~20 times faster texture construction
-                SystemInfo.IsFormatSupported(GraphicsFormat.R8G8B8_UNorm, FormatUsage.Sample)
-                    ? GraphicsFormat.R8G8B8_UNorm
-                    : GraphicsFormat.R8G8B8A8_UNorm,
-                TextureCreationFlags.DontInitializePixels | TextureCreationFlags.DontUploadUponCreate
-#endif
             );
             exportTexture.ReadPixels(new Rect(0, 0, destRenderTexture.width, destRenderTexture.height), 0, 0);
             RenderTexture.ReleaseTemporary(destRenderTexture);