+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/ImageConversionImageLoader.cs
@@ -13,16 +13,6 @@ using UnityEngine;
 using UnityEngine.Experimental.Rendering;
 using UnityEngine.Profiling;
 
-#if UNITY_6000_0_OR_NEWER
-using GLTFast.Loading;
-#else
-using System.Runtime.InteropServices;
-using GLTFast.Jobs;
-using GLTFast.Loading;
-using Unity.Collections.LowLevel.Unsafe;
-using Unity.Jobs;
-#endif
-
 namespace GLTFast
 {
     static class ImageConversionImageLoader
@@ -68,20 +58,6 @@ namespace GLTFast
             CancellationToken cancellationToken
         )
         {
-#if !UNITY_6000_0_OR_NEWER
-            var managedData = new byte[data.Length];
-            var gcHandle = GCHandle.Alloc(managedData, GCHandleType.Pinned);
-            var job = CreateMemCopyJob(data, gcHandle);
-            var jobHandle = job.Schedule();
-            while (!jobHandle.IsCompleted)
-            {
-                await Task.Yield();
-            }
-            jobHandle.Complete();
-            gcHandle.Free();
-            if (cancellationToken.IsCancellationRequested)
-                return ImageResult.Null;
-#endif
             while (context.DeferAgent.ShouldDefer())
             {
                 if (cancellationToken.IsCancellationRequested)
@@ -96,14 +72,7 @@ namespace GLTFast
                 settings.GenerateMipMaps,
                 settings.AnisotropicFilterLevel
                 );
-            texture.LoadImage(
-#if UNITY_6000_0_OR_NEWER
-                data.AsReadOnlySpan(),
-#else
-                managedData,
-#endif
-                !readable
-            );
+            texture.LoadImage(data.AsReadOnlySpan(), !readable);
             Profiler.EndSample();
             return new ImageResult(texture);
         }
@@ -131,23 +100,6 @@ namespace GLTFast
             };
             return txt;
         }
-
-#if !UNITY_6000_0_OR_NEWER
-        static unsafe MemCopyJob CreateMemCopyJob(
-            NativeArray<byte>.ReadOnly data,
-            GCHandle gcHandle
-        )
-        {
-            var job = new MemCopyJob
-            {
-                bufferSize = data.Length,
-                input = (byte*)data.GetUnsafeReadOnlyPtr(),
-                result = (void*)gcHandle.AddrOfPinnedObject()
-            };
-
-            return job;
-        }
-#endif // !UNITY_6000_0_OR_NEWER
 
         /// <summary>
         /// UnityWebRequestTexture always loads Jpegs/PNGs in sRGB color space