+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/VertexBufferBones.cs
@@ -176,11 +176,7 @@ namespace GLTFast
                     jobTangentI.input = (byte*)input;
                     jobTangentI.outputByteStride = outputByteStride;
                     jobTangentI.result = output;
-#if UNITY_COLLECTIONS
                     jobHandle = jobTangentI.ScheduleBatch(count, GltfImport.DefaultBatchCount);
-#else
-                    jobHandle = jobTangentI.Schedule(count, GltfImport.DefaultBatchCount);
-#endif
                     break;
                 case GltfComponentType.UnsignedShort:
                 {
@@ -191,11 +187,7 @@ namespace GLTFast
                         outputByteStride = outputByteStride,
                         result = output
                     };
-#if UNITY_COLLECTIONS
                     jobHandle = job.ScheduleBatch(count, GltfImport.DefaultBatchCount);
-#else
-                    jobHandle = job.Schedule(count, GltfImport.DefaultBatchCount);
-#endif
                     break;
                 }
                 case GltfComponentType.UnsignedByte:
@@ -207,11 +199,7 @@ namespace GLTFast
                         outputByteStride = outputByteStride,
                         result = output
                     };
-#if UNITY_COLLECTIONS
                     jobHandle = job.ScheduleBatch(count, GltfImport.DefaultBatchCount);
-#else
-                    jobHandle = job.Schedule(count, GltfImport.DefaultBatchCount);
-#endif
                     break;
                 }
                 default: